+++ config/filesystems.php
@@ -47,6 +47,13 @@ return [
             'report' => false,
         ],
 
+        'forms_output' => [
+            'driver' => 'local',
+            'root' => storage_path('app/forms_output'),
+            'visibility' => 'private',
+            'throw' => false,
+        ],
+
         's3' => [
             'driver' => 's3',
             'key' => env('AWS_ACCESS_KEY_ID'),